# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a NodeJS Boat project to demonstrate Swagger Documentation",
        "version": "1.0.0",
        "title": "Boats API",
        "contact": {
            "email": "dipakdip33@gmail.com"
        }
    },
    "schemes": ["http"],
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/allboatimage": {
            "get": {
                "summary": "Get all the image url from owner's boat",
                "description": "Get all the image url from owner's boat",
                "produces": ["application/json"],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/todosResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/ownerboatimage": {
            "get": {
                "summary": "get owner's boat image url",
                "description": "get owner's boat image url",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "req body",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "ownerId": {
                                "type": "object",
                                "$ref": "#/definitions/ownerId"
                            }
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/todosResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
        "/userLogin": {
            "post": {
                "summary": "User's Login",
                "description": "User's Login",
                "produces": ["application/json"],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Required Parameter",
                    "required": true,
                    "schema": {
                        "type": "object",
                        "properties": {
                            "task": {
                                "type": "object",
                                "$ref": "#/definitions/login"
                            }
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/todosResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/test": {
            "get": {
                "summary": "Test api",
                "description": "Test api",
                "produces": ["application/json"],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/todosResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "todosResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "task": {
                    "type": "string"
                },
                "assignee": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "ownerId": {
            "type": "object",
            "properties": {
                "ownerId": {
                    "type": "string"
                }
            }
        },
...
        "login": {
            "type": "object",
            "properties": {
                "Email": {
                    "type": "string"
                },
                "Password": {
                    "type": "string"
                }
            }
        },
...
        "InvalidResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }

        }
    }
}